+++ src/main/java/js/footballclubmng/service/NewsService.java
@@ -1,12 +1,12 @@
 package js.footballclubmng.service;
 
-import js.footballclubmng.dto.NewsDto;
+import js.footballclubmng.model.dto.NewsDto;
 import js.footballclubmng.entity.News;
 
 import java.util.List;
 import java.util.Optional;
 
 public interface NewsService {
-    Optional<News> getNewsById(int id);
+    Optional<News> getNewsById(long id);
     public List<NewsDto> findAllNews();
 }
